Swarm batch 8: Weddings::Tick and JukeBoxController::Add byte-exact
make verify 488/491 -> 490/492; sheet byte-exact 1587 -> 1602.

Weddings::Tick (1071 instr) converged. Two dead-statement forms are required for
byte fidelity: each step-3 `if ((*it)->countdown == N && BothPresent(self, *it))`
needs an `else { (*it)->countdown == 1; }` (a dead comparison) which bcc32 lowers
to the reference's jmp-skipped `sete dl; and edx,1` blocks, and the step-0
`if (!BothPresent(...)) { ...; break; }` needs an `else` setting step/countdown to
emit the reference's extra dead jmp. Fragile: a future "cleanup" would break it.

JukeBoxController::Add converged -- the real cause was JukeBox's constructor
parameter width: bcc32 emits a 16-bit load for a `short`/`unsigned short` ctor
parameter regardless of the caller's argument, and only an `int` gives the
reference's dword push. Jukebox.h/.cpp changed short -> int; the ctor stays
byte-exact.

Npccontrol: NpcControl_Tick is now implemented in full (348 lines) but is 84
hunks off (frame -240 vs -236; the aggro/chase block and the broadcast EH marker
swaps). Its real body is in the tree with the placeholder retained; residual
notes are in the agent report.

Mapcontrol: FUN_00482834 was attempted and reverted. The blocker is now precise:
the reference declares ~50 named `AnsiString` temporaries (one per SubString),
and inlining them bloats the frame by 36 bytes (-656 vs -620), which shifts every
[ebp-N] and ~93% of the instructions. Match the named-temp declarations.

Packets: Message_BuildServerStatus is 19 instructions from exact; the only
residual is the branch-end destructor idiom (three scope arms around a final
`*out += temp; return out;`) that also blocks Server_BuildOnlineNames and
Party_EncodeMemberList. `+=` is required; the arms are not produced by any form
tried. Worth a dedicated probe.

+++ src/Jukebox.cpp
@@ -5,7 +5,7 @@
 
 #pragma package(smart_init)
 
-JukeBox::JukeBox(short map_id)
+JukeBox::JukeBox(int map_id)
 {
     this->map_id = map_id;
     active = 0;
